requirements.txt
discord.py
psycopg2